Fix: ensure we include the right Lua header files
#### Brief overview of PR changes/additions
For Windows builds modify the `#include` lines for Lua header files to
specify the 5.1 version.

#### Motivation for adding to Mudlet
The default version - and the one needed for some packages like Luarocks
is a 5.4 one - and that includes header files in the "default" `include`
directory. So the headers that get pulled in are the wrong ones, which fail
to work as they are not compatible with Lua 5.1; to get the 5.1 instead I
believe we need to explicitly include the version specific sub-directory in
the `#include` lines.

#### Other info (issues closed, discussion etc)
This should be simpler to do than what is being attempted by #7841.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>