Add: sysConsoleSizeChanged event triggered on resize and timestamps toggling
#### Brief overview of PR changes/additions
Adds functions to enable/disable and find out the state of timestamps on
all user accessible windows. Also adds a `sysConsoleSizeChanged` event that
reports changes in the number of columns, rows and rows occupied by the
time-stamp (which is now localiseable for the end-users UI) for any user
window as well as the main Window.

#### Motivation for adding to Mudlet
Allow UI designers to accommodate the space used for timestamps. Also allow
time-stamps to be enabled on other console besides the main one if the end-
user desired. This is done with three new Lua API functions:
* `disableTimeStamps([windowName])` - turns off the timestamps if they are on
for the main (if no windowName is provided or it is `"main"`, or the
specified subconsole or user windows if it exists) - or returns a `nil` +
error message if they are already turned off.
* `enableTimeStamps([windowName])` - turns on the timestamps if they are off
for the main (if no windowName is provided or it is `"main"`, or the
specified subconsole or user windows if it exists) - or returns a `nil` +
error message if they are already turned on.
* `timeStampsEnabled([windowName]) - returns true or false to indicate for
the main (if no windowName is provided or it is `"main"`, or the
specified subconsole or user windows if it exists) whether timestamps are
shown for the particular console.

#### Other info (issues closed, discussion etc)
These additions were prompted by the Discord user **missionz3r0** in our
mudlet-development channel on 2025/05/20 04:30 UTC
https://www.discord.com/channels/283581582550237184/283582439002210305/1374242960299786352
> Gonna add  this to things I'd like to fix/improve when I'm no longer
> moving houses/dead tired.
>
> **Context,**
> As far as I am aware, there is no way to know if the main console is
> showing timestamps, nor is there a way to listen in on if they are
> toggled.  If a package dev tries to borders such that the main console is
> exactly the size of the wrap_at * font-size, timestamps will cause the
> window to overflow.
>
> **What I'd like to do,**
>* Add a lua function that'll return boolean representing the state of the
> timestamp toggle,
>* Raise a `sysSettingChanged` event when timestamps are toggled,
>* Include timestamp characters when calculating `getMainConsoleWidth`

In implementing these changes I realised that there was some duplication
of code for each `TTextEdit` instance in a `TConsole` which was better done
in the parent `TConsole`; it also meant the "slot" that processed the
time-stamp button was moved from the `TTextEdit`s to the `TConsole` so that
it could also update the relevant toolbutton in the bottom button
collection for the `TMainConsole` instance.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>