Infrastructure: adjust name of TConsole::setFont(...)
Renames `(bool) TConsole::setFont(const QString&)` to `setFontName`
because it does actually do the latter (stores the NAME of the font to use)
and could otherwise be mistaken for the base class
`(void) QWidget::setFont(const QFont&)` method.

To reduce the chance of misinterpreting what the method does in comparision
with what the base class function performs.

It is not impossible that we might restructure things so that we send a
reference to nn actual font instance rather than the name of a font in this
class. Such a `setFont` method then would then have the `override` "label"
applied to it so as to flag that it is replacing the base class function,
which the method renamed in this PR does **not**.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>