Infrastructure: update Windows Github runner image used
Change from the `windows-2019` to the `windows-2022` image for Windows GHA
runs of our CI/CB process.

The oldest image we were using is, as of 2025/06/01, now deprecated and
will be removed shortly; as it is, it is being browned out for part of
today! See: https://github.com/actions/runner-images/issues/12045

Unlike the removal of the oldest Linux image this should not be too much
of a faff to sort out...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>