Infrastructure: rationalise font handling
\#### Brief overview of PR changes/additions
* Revert redesign of main font selection in preferences.
* Restructure usage of fonts to remove the somewhat redundent `(QFont)
mDisplayFont` for widgets that inherently have their own (QFont) which they
use for painting operations.
* Provide a means to track only the details of `QFont`s that we care about
which should be more lightweight than holding a complete copy of a font (in
`TFontDetails`).
* Include the "antialisaing" detall for the main display font (only for the
main console) in the active updates that changing the main font in the
preferences produces.
* Apply changes to the font in a `TConsole` to **all** the `TCommandLine`s
associated with it.

\#### Brief overview of PR changes/additions
* The previously revised font settings in the preferences that used a
native font selection was not as useful as it seemed - at least the Windows
one offered controls that we did not need or use which meant they had not
actual effects which would be confusing to the end user.
* Storing a separate copy of `QFont`s did not seem productive especially as
it was easy to modify the intrinsic one and not the copy or vice-versa.
* Adding a separate class to track the font details we ARE interested in
makes it easier to compare fonts and to tell whether two instances are the
same as far as Mudlet is concerned
* It was not clear that the "anti-aliasing" setting was being correctly
applied/used where it was intended - with this PR it is also updated as it
is changed in the prefernces along with the other two settings: font
"family" and "size".
* The font in command-lines now clearly follow the main console or a sub-
console that they belong to.

\#### Other info (issues closed, discussion etc)
To be filled out

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>