fix: broken AND-Triggers with only one item
PPrevent a trigger with only one "item" from being
recored as a multi-line (AND) trigger. The prior
code meant that after setting the multi-line delta
to a value greater than `-1` which now flags that
a trigger is of the AND rather than the OR type
ignored the later deletion of all but one of the
"Items" with the result that later on the trigger
failed to operate correctly.

Prevent a stupid odditiy in the behavour which
borked the trigger concerned.

This was brought to my attention by a Discord
user who kindly also raised the issued which this
PR should close #7894.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>