Improve: allow IP Version selection for Game Server connection
#### Brief overview of PR changes/additions
Revise the `QSslSocket` connection process in `cTelnet` class to split it
up into "Host Lookup" + "Making Connection" + "Encryption negotiation" -
this is so that the first two can be made common with the insecure
connection process - and so that Mudlet can examine the Host lookup results
in the same manner for both. The prior use of
`QSslSocket::connectToHostEncrypted(...)` combined all three of those steps
but meant there was no ability to examine the results of the Host Lookup.
As a result the messaging can be made more uniform for both cases and
detail the lookup results better - as well as allowing for selection of
IPv4 or 6 only as needed.

#### Motivation for adding to Mudlet
@Kebap has been having problems recently with a bogus IPv6 lookup result
and this should help with that. This should help - but not solve - #3053
and #3791 - and hopefully it can allow me or someone else - to develop a
"happy eye-balls" approach to try opening the first valid IPv4 and IPv6
addresses at the same time. However that will need further work as it will
require the handling of multiple `QTcpSocket`/`QSslSocket`s simultaneously
and the current `cTelnet` class will need some serious reorganising to do
this.

#### Other info (issues closed, discussion etc)
* Reorders a couple of methods declared in the `Host.h` file - they were
out of position.
* Renames:
  * `(std::unique_ptr<QNetworkProxy>) Host::mpDownloaderProxy`
                                              ==> `Host::mpDownloaderProxy`
  * `(QSslSocket/QTcpSocket) cTelnet::socket` ==> `cTelnet::mpSocket`
  * `(QString) cTelnet::hostName` ==> `cTelnet::mHostName`
  * `(int) cTelnet::hostPort` ==> `cTelnet::mHostPort`
* Restores reporting of non-SSL related `QTcpSocket`/`QSslSocket` errors -
except for the one associated with the Server disconnecting (usually by
`quit`-ting from a MUD)!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>