Add: load profiles from command line
Adds a command line `-l` / `--load` option to go alongside the `-p` /
`--profile` option to specify profiles to be loaded but NOT connected on
startup. Similarlarly add handling for a colon delimited list as
`MUDLET_LOAD_PROFILES` to do the same thing via environmental variables.
Modify existing handling of `MUDLET_PROFILES` to also consider
`MUDLET_CONNECT_PROFILES` as a (recommended) alternative to give profiles
to connect on start-up.

A recent discussion on Discord (between myself, **bazzarranti** and
**atari2600tim** in the **Help** channel pointed out that it was not
possible to "load" profiles from the "commnadline" although it was possible
to "connect" them.

> bazarranti — 01/08/2025 21:37 UTC
> I would like to be able to use "open profile on mudlet start" or open a
> profile from the command line, but open it in offline mode. I can't seem
> to figure out how to make that happen though. Can anyone help?

This maintains the order of profiles given, handling command-line options
before environmental ones - and not ignoring the latter if the former is
used. It also handles the same profile being requested twice and also
conflicting requests to both load and connect a profile (the connect option
will always win) - it also reportes (to the OS command line) a profile that
is not recognised.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>