Add UI-design-philosophy.md (#8002)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Document the UI design philosophy that has been guiding Mudlet's UI
choices for the past 16 years to the success it has today.
#### Motivation for adding to Mudlet
So that everyone working on Mudlet's interface is going in the same
direction, and the interface is consistent as a result.
#### Other info (issues closed, discussion etc)
Input on this is welcome, especially from UI design professionals
familiar with MUDs.