Infrastructure: move from squirrel.windows to Clowd.squirrel (PR FORM)
#### Brief overview of PR changes/additions
Replaces the original squirrel.windows Windows installer builder with
the Clowd.squirrel one which although now "archived" (and thus static) does
produce working installers for our Windows builds (which WILL install all
the other libraries that Mudlet needs to run - the orginal one now fails
to do so and has not being updated for several years).

#### Motivation for adding to Mudlet
To get Windows PTBs (and hopefully Releases) installers working again.

#### Other info (issues closed, discussion etc)
This should contain the same code as:
https://github.com/Mudlet/Mudlet/tree/Infrastructure_moveFrom_squirrel.windows_to_Clowd.squirrel
which exists to test the actual Windows PTB building capability

However this PR also needs to compile (to prove that the changes do not
bork the "Snapshot" builds either).

Changes to either branch should be duplicated in the other and they both
need to work (apart from the registering the build with DBLSQD of the other
as that relies on a separate workflow to create the feed needed).

It refactrors the script that generates "artifacts" for the Windows builds
so that as well as the single executable that is zipped up and then
uploaded to the Mudlet website where it is unzipped it also generates an
(zip) archive of the (signed) files that are in the "installer" prior to
them being incorporated in the installer - this actually reproduces what
is done for "Snapshot/PR/Testing" builds which does that but uploads the
archive to the Mudlet website - instead the archive is retains as a GH
artefact for the default (90?) days. Also the compression used is maximised
for both the archives containing the built files - and disabled for the
single "installer" artefact as that is already compressed.

**This has NOT been tested for "Release" type builds - and that might be
problematic at the time the next release is attempted!**

This is a /claim #7687

It may also /claim #7970 - but that can only be proved by testing when the
code is merged - and might take a couple of days to varify as testing the
main repo branch version does point out that with over 100 commits the
change log is too large to include.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>