Infrastructure: get link-ptbs-to-dblsqd workflow working again
#### Brief overview of PR changes/additions
Remove use of `ARCH` (enviromental) variable from Windows CI/CB system
and hard code the use of `x86_64` everywhere it was used as we not longer
produce 32-Bit Windows builds. This change had previously impacted the
named workflow as that was looking for the 32-bit Windows releases that we
were no longer reporting to DBLSQD.

#### Motivation for adding to Mudlet
Get PTBs produced properly again.

#### Other info (issues closed, discussion etc)
After #8015 this should finally fix and so /claim #7970.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>