Improve: move Ambiguous East Asian Width option control from preferences to setConfig() (#8061)
#### Brief overview of PR changes/additions
A replacement for #8020 that actually works - rather than removing the
ability to control the "Ambiguous East Asian Width Characters"
completely it just removes the control from the preferences and, unlike
that failed PR, moves it to something that the Lua API (and thus the
end-user) can examine and control via the `getConfig(...)` and
`setConfig(...)` interface. The setting is called
`ambiguousEAsianWidthCharacters` and the three values it can have are:
* `"narrow"` (with the same effect as the removed checkbox being
"unchecked")
* `"wide"` (same effect as "checked")
* `"auto"` (same effect as "partially-checked")

It also synchronises the `getConfig()` function (which takes no
arguments and returns all the settings it knows about in a table) in the
`Other.lua` external file. There were several settings that had been
added to the internal C core code but which hadn't been added to the
external Lua function that supplements it to handle a "no-arguments"
usage. Whilst those two are now in step this PR does not make any
guarantees about synchronicity with the C core code `setConfig(...)`
function!

#### Motivation for adding to Mudlet
To remove a not often needed control from the preferences to simplify
them.

#### Other info (issues closed, discussion etc)
Whilst correctly updating the "bulk" `getConfig(...)` in the `Other.lua`
file to support this new (there) setting I found a number of others that
had been added to the C++ core `getConfig(...)` function but which had
not been added in that external Lua file so which would be missed if the
user tried to get ALL the settings available.

This replaces/supplants PR #8058 which can be closed as not required.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>