Fix: Linux 'light' apperance still keeping Mudlet dark when OS theme is dark as well (#8098)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Setting the 'Light' theme in Mudlet when the overall OS theme was dark
still kept Mudlet in the dark mode - this fixes it.

Also, removed our custom dark theme since Qt supports it natively now.
#### Motivation for adding to Mudlet
Bugfix and the less custom code we have without a good reason, the
better.
#### Other info (issues closed, discussion etc)
This requires Qt 6.5, so https://github.com/Mudlet/Mudlet/pull/8081
needs to go in first.

---------

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>