Improve: add instanced rendering to 3D mapper (#8140)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Implements [instanced
rendering](https://learnopengl.com/advanced-opengl/instancing), so
instead of asking OpenGL to draw 100 rooms 100 times, we ask it 1 time
to draw 100 rooms. Due to the CPU -> GPU communication, it is the
**ask** itself that's expensive, so this makes things a lot quicker by
drawing a lot more things at once.
#### Motivation for adding to Mudlet
Improved performance for the 3D mapper
#### Other info (issues closed, discussion etc)
Because things need to be drawn in the right order in OpenGL, batching
all of the calls into one call might have side-effects, need to make
sure there's no regressions.

---------

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>