Improve: Allow users to specify saveProfile filename (#7982)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Make the lua function `saveProfile(path)` save AS filename.xml if the
path ends with .xml

#### Motivation for adding to Mudlet
I was trying to make a function to store my profile in a separate folder
and there didn't seem to be a good way to avoid stacking up duplicate
copies. By adding this functionality it becomes simple to automatically
overwrite the old copy.

#### Other info (issues closed, discussion etc)

---------

Co-authored-by: Vadim Peretokin <vperetokin@hey.com>
Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>