Infrastructure: update files listed in qmake project file
#### Brief overview of PR changes/additions
Include in the `./src/mudlet.pro` qmake project file the details of some
files that are either recently added - or not previously included.

#### Motivation for adding to Mudlet
So that those of us still using Qt Creator in "qmake" mode can see project
files - and so that they are indexed within that application for when
searches are done "Project wide". Without this, usages of things can be
missed when editing is being done.

#### Other info (issues closed, discussion etc)
This should not provoke any functional changes to the Mudlet application!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>