Fix: switch to FFMPEG audio backend on Windows
#### Brief overview of PR changes/additions
Change from using the "Window Media File (WMF)" audio backend to the
"FFMPEG" one by default (can still be overridden by setting the
`QT_MEDIA_BACKEND` variable to `windows`). Ammends the CI/CB scripts to
deduce the needed libraries to package for all the Qt plugins as well as
the main `Mudlet.exe` executable automatically - does away with a manual
method - which, given the number of additional libaries that the FFMPEG
backend requires - was fragile and would likely break on MSYS2+Mingw-w64
updates.

#### Motivation for adding to Mudlet
To allow Mudlet to play `.ogg` and `.opus` media files on Windows. The
previously used `windows` backend does not support those audio formats -
which renders them inaudiable in MUDs that use them (like the "Last
Outpost" MUD).

#### Other info (issues closed, discussion etc)
This PR incorporates some ideas originally posted in @mpconley 's #7955
which also tried to do the same thing but also tweaked audio playback
on other OSes and included the production of a lot of debugging information
which is probably not required. Using the MSYS2+Mingw-w64 packages the
additional library requirements/dependencies will automagically be
satified by the installation of the base `ffmpeg` and `qt6-ffmpeg` packages
so there is little or no need to manually regorously check for them.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>