Add: getKeyCode function to return key code and modifier of a key-binding
#### Brief overview of PR changes/additions
Adds `getKeyCode(keyBindingName/ID)` and returns two integer values, the
key "code" and the "modifiers" if any for the specified keybinding.

This, used in conjuction with `findItems("", "keybinding", false)` to get
the ID numbers of all extant key-binding will allow a script/package/end-
user to determine whether a particular key has already had a binding
assigned to it to avoid unwanted multiples bindings to the same key or key
and modifier.

#### Motivation for adding to Mudlet
A discussion initiated at 2025/10/22 06:52 UTC in Mudlet's Discord
[**#mudlet-development**](https://www.discord.com/channels/283581582550237184/283582439002210305/1430448688831926404) channel by the new user **@zygmud
**.
> can I ask a small question here?
>
> I would like to check if a hotkey exists, without knowing its ID or name.
Just check if a keybind exists on a given key, like mudlet.key.F12, for
example, is this possible?
> I'm writing a script for others to import, and I don't want to overwrite
> existing settings

#### Other info (issues closed, discussion etc)
The numbers returned are the same as defined in the `mudlet.key` and
`mudlet.keymodifier``tables as listed in the
`.\src\mudlet-lua\lua\KeyCodes.lua` file in the source code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>