Improve: make Windows "portable" builds self-extracting archives
#### Brief overview of PR changes/additions
Revises the Windows OS CI/CB process to:
* Produce a "portable" version of Mudlet for PTBs as well as "Release"
versions
* Exit cleanly earlier in the process for when a PTB is NOT required
{via ABORT_WORKFLOW environmental variable} when no changes to the
`development` branch happened in the last 24 hours.)
* Identify "what" the CI/CB process is being used for once, early on, and
forwarding that to later stages {via BUILD_ACTION environmental variable}.
I.e. adopting a DRY approach!
* Eliminating a couple of environmental variables as they duplicate ones
that MSYS2+Mingw-w64 already provides:
  * `"mingw-w64-${BUILDCOMPONENT}"` is the same as `"${MINGW_PACKAGE_PREFIX}`
  * `${MINGW_BASE_DIR}` was a Windows (typically `C:/msys64/mingw64`) path
  that can be replaced by "${MINGW_PREFIX}" in a POSIX form (typically
  `/mingw64`) as far as Qt Creator/`qmake` is concerned.
* Fix a defect in the current `7zip` package which cannot make self-
  extracting archives without the "self-extracting" module (`7zCon.sfx`)
  being placed alongside the executable (the MINGW-W64 people think it
  should be placed  in a `./lib/7zip` directory but the executable cannot
  see it there - unless it gets patched - which is a bit beyond me whereas
  putting it next to the executable isn't:
  https://github.com/msys2/MINGW-packages/pull/25876 )!
* Revise some elements of the build naming so that builds that did contain
  `-testing-pr####` lose the `-testing` prefix - so that that is ONLY used
  for CI/CB builds when a PR is *merged* into the `development` branch.
  This should clarify the naming on the web-page that carry the
  "snap-shots"!

#### Motivation for adding to Mudlet
Improve end-user experience for those wanting a "Portable" Windows build;
a file that they can run to produce all the files needed for it instead of
having to use a utility (or relying on Windows own zip file handling) to
unzip the archive is simpler. Also,by making the archive a self-extrcting
one it will make it possible to update to a new version by triggering an
older version to run a downloaded copy of a newer one. Furthermore the code
and scripts in this PR should enable the production of "portable" PT Builds
(and PR or "Testing" builds) besides the "Release" only one that the prior
situation was. It also signs the files and the self-extracting archive in
the case of PTB and Release builds which should mean less problems with
anti-virus systems.

#### Other info (issues closed, discussion etc)
This is a necessary prequel to making "portable" PTB and Release builds be
capable of being updated. With the prior scripts only configure to produce
a portable build for "Release" type builds - and there not having been one
of those since the changes for portable builds were put in place the usage
of such a "portable" mode of operation may not have been as extensively
tested as we might like; hopefully this PR will help to change that!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>