Fix: use more HTML entities in translatable messages
#### Brief overview of PR changes/additions
In #8145 an attempt was made to simplify translations by removing mulitple
uses of the first replaceable parameter for a non-breaking space. This is
necessary because Qt's tools - such as the Qt Creator IDE - used to (and
may still) silently convert raw-string versions to ordinary spaces. However
it was felt that this was confusing for translators and it was decided to
replace those with Unicode escape codes (`\uXXXX`) instead. These do get
converted to the (X|HT)ML equivalents (`&xXXXX;`) when the relevant files
are seen in CrowdId. Anyhow there is also some usage of non-breaking
hyphens (`\u2011`) in the same texts which could be confused with ordinary
hyphens visually and get shown as seemingly Mojibake in some other
situations (e.g. `gitk` on Windows!)

#### Motivation for adding to Mudlet
Further improve on implementation of #1802.

#### Other info (issues closed, discussion etc)
I do wonder if we are going up a blind alley with this - originally the
translations were filled with `%1`s {the replacement text is allowed to
be used more than once in a translation so the concept of one `.arg(...)`
to one `%#` is incorrect) but now those are replaced with longer (but not
more legible IMHO) `\u0A` which get converted to `&#0A;` at CrowdIn.

:man_shrugging:

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>