Infrastructure: improve build type handling
#### Brief overview of PR changes/additions
Codifies the "type" of the Mudlet build to a single `enum class` instead
of several `bool` flags and removes duplicate code to determine what
value(s) they/it should have.

#### Motivation for adding to Mudlet
Simplification and removal of duplicate code. Also preparation wanted as
part of solving problem with updating current Windows builds with the CI/CB
produced "installer".

#### Other info (issues closed, discussion etc)
This is the first step in the resolution I am attempting for #7313.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>