Update: revise displayed Copyright year to 2026
#### Brief overview of PR changes/additions
Change places in text where the current year is display as part of the
Copyright information.

#### Motivation for removing from Mudlet
We do this manually so as to avoid offending those looking for
"reproducable builds" - and thus we do not define the year as a value
determinated at compilation time (typically by manipulation of the __DATE__
value)!

#### Other info (issues closed, discussion etc)
The Windows part of the installers repository ALSO needs a single change.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>